feat: refactor comms to functional style
