+++ src/services/storage/storage.processor.ts
@@ -6,11 +6,10 @@ import stake from './handlers/stake'
 import { Handler } from '../../definitions'
 import { Eth } from 'web3-eth'
 import { StorageServices } from './index'
-import { Comms } from '../../communication'
 
 const HANDLERS: Handler<StorageServices>[] = [offer, request, stake]
 
-export default function (services: StorageServices, deps: { eth?: Eth, comms?: Comms }) {
+export default function (services: StorageServices, deps: { eth?: Eth }) {
   return async (event: EventData): Promise<void> => {
     const promises = HANDLERS
       .filter(handler => handler.events.includes(event.event))
